+++ lib/business.ts
@@ -1,0 +1,7 @@
+export const business = {
+  name: "Rock of Ages Auto Repairs",
+  type: "Mobile Mechanic",
+  countryCode: "+44",
+  mobile: "7845465922",
+  email: "catalinrol@gmail.com"
+};
