+++ lib/content.ts
@@ -1,0 +1,7 @@
+import { business } from "./business";
+
+export const homePageBusinessIntro = `Hey there! Welcome to your local mobile mechanic business, where we're all about bringing convenience and top-notch service right to your doorstep! At ${business.name}, we understand that life can get busy, and car maintenance is often the last thing on your mind. That's why we've made it our mission to make taking care of your vehicle as easy and stress-free as possible.
+Picture this: you're at home enjoying your day off, and suddenly you realize your car is due for an oil change. No need to fret or rearrange your plans - just give us a call, and we'll be there as soon as we can! Our team of experienced mobile mechanics specializes in oil service and brake replacement. Need a coolant or brake fluid replacement, or perhaps new spark plugs? We've got you covered! With our fully equipped mobile service vehicle, we'll have your car running smoothly in no time, without you ever having to leave the comfort of your home.
+But here's the best part - we're not just here to fix your car; we're here to make your life easier. Our friendly and reliable mobile mechanics are committed to providing exceptional service with a smile. We take great pride in our work and ensure that each vehicle receives care and attention, inspiring trust and confidence from our customers. So, why stress over car maintenance when you can leave it to the pros? Sit back, relax, and let ${business.name} take care of all your automotive needs - because when it comes to keeping your car in tip-top shape, we've got you covered.`;
+
+
